Move logger to separate file
