+++ visualizer.py
@@ -25,43 +25,8 @@ import webinterface as web_mod
 import asyncio
 import atexit
 from waitress import serve
-import traceback
 
-import logging
-from logging.handlers import RotatingFileHandler
-
-# Create a custom logger
-logger = logging.getLogger(__name__)
-
-# Set the level of this logger.
-logger.setLevel(logging.DEBUG)
-
-# Create handlers
-console_handler = logging.StreamHandler()
-file_handler = RotatingFileHandler('visualizer.log', maxBytes=500000, backupCount=10)  # adjust as needed
-
-# Set the level for handlers
-console_handler.setLevel(logging.DEBUG)
-file_handler.setLevel(logging.DEBUG)
-
-# Create formatters and add it to handlers
-formatter = logging.Formatter('[%(asctime)s] %(levelname)s - %(message)s',
-                              datefmt='%Y-%m-%d %H:%M:%S')
-console_handler.setFormatter(formatter)
-file_handler.setFormatter(formatter)
-
-# Add handlers to the logger
-logger.addHandler(console_handler)
-logger.addHandler(file_handler)
-
-
-# Custom exception handler to log unhandled exceptions
-def log_unhandled_exception(exc_type, exc_value, exc_traceback):
-    logger.error("Unhandled Exception: ", exc_info=(exc_type, exc_value, exc_traceback))
-
-
-sys.excepthook = log_unhandled_exception
-
+from lib.log_setup import logger
 
 
 os.chdir(sys.path[0])
